requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.3
certifi==2020.6.20
chardet==3.0.4
discord.py==1.3.4
idna==2.10
multidict==4.7.6
praw==7.1.0
prawcore==1.4.0
pycodestyle==2.6.0
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
update-checker==0.17
urllib3==1.25.10
websocket-client==0.57.0
websockets==8.1
yarl==1.5.0